Fixed bug with postgree DateTime column
