+++ SunEngine/Controllers/CategoriesController.cs
@@ -37,9 +37,6 @@ namespace SunEngine.Controllers
 
         private CategoryInfoWithAccesses CategoryInfoWithAccessesFromCategory(Category category)
         {
-            // TODO сделать самостоятельную функцию которая проходит от корня 
-            // TODO идя сначала по ближайшим к корню нодам и потом ниже
-
             if (!authorizationService.HasAccess(User.UserGroups, category,
                     OperationKeys.MaterialAndMessagesRead) && category.Id != categoriesStore.RootCategory.Id)
             {
