+++ SunEngine/Presenters/BlogPresenter.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Linq;
 using System.Threading.Tasks;
 using SunEngine.Commons.DataBase;
@@ -13,7 +14,7 @@ namespace SunEngine.Presenters
         {
         }
 
-        public Task<IPagedList<PostViewModel>> GetPostsAsync(int categoryId, int page,int pageSize)
+        public Task<IPagedList<PostViewModel>> GetPostsAsync(int categoryId, int page, int pageSize)
         {
             return db.MaterialsNotDeleted.GetPagedListAsync(
                 x => new PostViewModel
@@ -34,6 +35,19 @@ namespace SunEngine.Presenters
                 page,
                 pageSize);
         }
-        
+    }
+
+    public class PostViewModel
+    {
+        public int Id { get; set; }
+        public string Title { get; set; }
+        public string AuthorName { get; set; }
+        public string AuthorLink { get; set; }
+        public string AuthorAvatar { get; set; }
+        public string Preview { get; set; }
+        public int MessagesCount { get; set; }
+        public DateTime PublishDate { get; set; }
+        public string CategoryName { get; set; }
+        public bool HasMoreText { get; set; }
     }
 }
